rent-a-printer/service/requirements.txt
pycups
pypdf[image]
qrcode[pil]
requests
flask
flask-login
pyipp
gunicorn
